Move `audit_plot` into `sync_fallback` module
