+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/benchmark.rs
@@ -11,7 +11,8 @@ use std::path::PathBuf;
 use subspace_core_primitives::crypto::kzg::{embedded_kzg_settings, Kzg};
 use subspace_core_primitives::{Record, SolutionRange};
 use subspace_erasure_coding::ErasureCoding;
-use subspace_farmer::single_disk_farm::farming::{audit_plot, PlotAuditOptions};
+use subspace_farmer::single_disk_farm::farming::sync_fallback::audit_plot;
+use subspace_farmer::single_disk_farm::farming::PlotAuditOptions;
 use subspace_farmer::single_disk_farm::{SingleDiskFarm, SingleDiskFarmSummary};
 use subspace_farmer_components::sector::sector_size;
 use subspace_proof_of_space::Table;
